Add faster version of faceblurring
