+++ faceblurring/utils.py
@@ -7,7 +7,9 @@ from tkinter import filedialog, simpledialog
 
 import cv2
 import numpy as np
-from colorama import Fore
+import onnxruntime as ort
+import torch
+from colorama import Fore, Style, init
 from scipy.special import expit
 
 
@@ -68,61 +70,6 @@ def gen_step_frames(vid_fps, step_vid_length):
 
 def make_out_name(outdir, part_id, vid_name, img_id):
     return os.path.join(outdir, f"{part_id}_{vid_name}_{img_id:05}.jpg")
-
-
-def blur_faces(frame, faces, debug=False):
-    if not len(faces):
-        return frame
-
-    frame_h, frame_w, _ = frame.shape
-
-    for face in faces:
-        x1, y1, x2, y2 = resize_box(face["bbox"], frame_h, frame_w)
-        conf = face["det_score"]
-
-        if debug:
-            # Include the rect and conf
-            cv2.rectangle(frame, (x1, y1), (x2, y2), (0, 0, 255), 2)
-            text = "{:.2f}".format(conf)
-            # Display the label at the top of the bounding box
-            label_size, base_line = cv2.getTextSize(
-                text, cv2.FONT_HERSHEY_SIMPLEX, 2, 1
-            )
-            top = max(y2, label_size[1])
-            cv2.putText(
-                frame,
-                text,
-                (x1, top - 4),
-                cv2.FONT_HERSHEY_SIMPLEX,
-                0.4,
-                (255, 255, 255),
-                1,
-            )
-
-        # Add the blurring in
-        roi = frame[y1:y2, x1:x2]
-
-        try:
-            # Blur the coloured image
-            blur = cv2.GaussianBlur(roi, (101, 101), 0)
-            # Insert the blurred section back into image
-            frame[y1:y2, x1:x2] = blur
-        except:
-            if debug:
-                print(f"[ERROR] Blurring failed.")
-
-    return frame
-
-
-def resize_box(box, frame_h, frame_w):
-    x1, y1, x2, y2 = (
-        max(int(box[0]), 0),
-        max(int(box[1]), 0),
-        min(int(box[2]), frame_w),
-        min(int(box[3]), frame_h),
-    )
-
-    return (x1, y1, x2, y2)
 
 
 def create_csv(output_dir_images, img_id, csv_path):
@@ -218,3 +165,13 @@ def tidy_up(vid_files, output_dir):
         except:
             print("Could not remove blurred timelapse video (it might still be open?)")
             input("Press enter to retry")
+
+
+def check_device():
+    if ort.get_device() != "GPU":
+        print(
+            f"{Fore.YELLOW}WARNING: {Style.RESET_ALL}Could not find GPU. Falling back to CPU."
+        )
+        return (480, 480), ["CPUExecutionProvider"]
+
+    return (640, 640), ["CUDAExecutionProvider"]
